+++ contrib/opencensus-ext-azure/tests/test_azure_metrics_exporter.py
@@ -18,10 +18,13 @@ from datetime import datetime
 import mock
 
 from opencensus.common import utils
-from opencensus.ext.azure import metrics_exporter
 from opencensus.ext.azure.common import Options
 from opencensus.ext.azure.common.protocol import DataPoint
-from opencensus.ext.azure.metrics_exporter import standard_metrics
+from opencensus.ext.azure.metrics_exporter import (
+    MetricsExporter,
+    new_metrics_exporter,
+    standard_metrics,
+)
 from opencensus.metrics import label_key, label_value
 from opencensus.metrics.export import (
     metric,
@@ -61,13 +64,13 @@ class TestAzureMetricsExporter(unittest.TestCase):
         instrumentation_key = Options._default.instrumentation_key
         Options._default.instrumentation_key = None
         self.assertRaises(ValueError,
-                          lambda: metrics_exporter.MetricsExporter())
+                          lambda: MetricsExporter())
         Options._default.instrumentation_key = instrumentation_key
 
     def test_constructor_invalid_batch_size(self):
         self.assertRaises(
             ValueError,
-            lambda: metrics_exporter.MetricsExporter(
+            lambda: MetricsExporter(
                 instrumentation_key='12345678-1234-5678-abcd-12345678abcd',
                 max_batch_size=-1
             ))
@@ -75,7 +78,7 @@ class TestAzureMetricsExporter(unittest.TestCase):
     @mock.patch('requests.post', return_value=mock.Mock())
     def test_export_metrics(self, requests_mock):
         metric = create_metric()
-        exporter = metrics_exporter.MetricsExporter(
+        exporter = MetricsExporter(
             instrumentation_key='12345678-1234-5678-abcd-12345678abcd')
         requests_mock.return_value.text = '{"itemsReceived":1,'\
                                           '"itemsAccepted":1,'\
@@ -90,7 +93,7 @@ class TestAzureMetricsExporter(unittest.TestCase):
 
     def test_export_metrics_histogram(self):
         metric = create_metric()
-        exporter = metrics_exporter.MetricsExporter(
+        exporter = MetricsExporter(
             instrumentation_key='12345678-1234-5678-abcd-12345678abcd')
         metric.descriptor._type = MetricDescriptorType.CUMULATIVE_DISTRIBUTION
 
@@ -98,7 +101,7 @@ class TestAzureMetricsExporter(unittest.TestCase):
 
     @mock.patch('requests.post', return_value=mock.Mock())
     def test_export_metrics_empty(self, requests_mock):
-        exporter = metrics_exporter.MetricsExporter(
+        exporter = MetricsExporter(
             instrumentation_key='12345678-1234-5678-abcd-12345678abcd')
         exporter.export_metrics([])
 
@@ -107,7 +110,7 @@ class TestAzureMetricsExporter(unittest.TestCase):
     @mock.patch('requests.post', return_value=mock.Mock())
     def test_export_metrics_full_batch(self, requests_mock):
         metric = create_metric()
-        exporter = metrics_exporter.MetricsExporter(
+        exporter = MetricsExporter(
             instrumentation_key='12345678-1234-5678-abcd-12345678abcd',
             max_batch_size=1)
         requests_mock.return_value.status_code = 200
@@ -123,7 +126,7 @@ class TestAzureMetricsExporter(unittest.TestCase):
 
     def test_create_data_points(self):
         metric = create_metric()
-        exporter = metrics_exporter.MetricsExporter(
+        exporter = MetricsExporter(
             instrumentation_key='12345678-1234-5678-abcd-12345678abcd'
         )
         data_points = exporter._create_data_points(metric.time_series[0],
@@ -138,7 +141,7 @@ class TestAzureMetricsExporter(unittest.TestCase):
 
     def test_create_properties(self):
         metric = create_metric()
-        exporter = metrics_exporter.MetricsExporter(
+        exporter = MetricsExporter(
             instrumentation_key='12345678-1234-5678-abcd-12345678abcd'
         )
         properties = exporter._create_properties(metric.time_series[0],
@@ -149,7 +152,7 @@ class TestAzureMetricsExporter(unittest.TestCase):
 
     def test_create_properties_none(self):
         metric = create_metric()
-        exporter = metrics_exporter.MetricsExporter(
+        exporter = MetricsExporter(
             instrumentation_key='12345678-1234-5678-abcd-12345678abcd'
         )
         metric.time_series[0].label_values[0]._value = None
@@ -161,7 +164,7 @@ class TestAzureMetricsExporter(unittest.TestCase):
 
     def test_create_envelope(self):
         metric = create_metric()
-        exporter = metrics_exporter.MetricsExporter(
+        exporter = MetricsExporter(
             instrumentation_key='12345678-1234-5678-abcd-12345678abcd'
         )
         value = metric.time_series[0].points[0].value.value
@@ -189,7 +192,7 @@ class TestAzureMetricsExporter(unittest.TestCase):
     def test_shutdown(self):
         mock_thread = mock.Mock()
         mock_storage = mock.Mock()
-        exporter = metrics_exporter.MetricsExporter(
+        exporter = MetricsExporter(
             instrumentation_key='12345678-1234-5678-abcd-12345678abcd'
         )
         exporter.exporter_thread = mock_thread
@@ -202,11 +205,10 @@ class TestAzureMetricsExporter(unittest.TestCase):
                 '.transport.get_exporter_thread')
     def test_new_metrics_exporter(self, exporter_mock):
         with mock.patch('opencensus.ext.azure.metrics_exporter'
-                        '.heartbeat_metrics.enable_heartbeat_metrics') as hb:
+                        '.statsbeat_metrics.collect_statsbeat_metrics') as hb:
             hb.return_value = None
             iKey = '12345678-1234-5678-abcd-12345678abcd'
-            exporter = metrics_exporter.new_metrics_exporter(
-                instrumentation_key=iKey)
+            exporter = new_metrics_exporter(instrumentation_key=iKey)
 
             self.assertEqual(exporter.options.instrumentation_key, iKey)
             self.assertEqual(len(exporter_mock.call_args_list), 1)
@@ -221,10 +223,10 @@ class TestAzureMetricsExporter(unittest.TestCase):
                 '.transport.get_exporter_thread')
     def test_new_metrics_exporter_no_standard_metrics(self, exporter_mock):
         with mock.patch('opencensus.ext.azure.metrics_exporter'
-                        '.heartbeat_metrics.enable_heartbeat_metrics') as hb:
+                        '.statsbeat_metrics.collect_statsbeat_metrics') as hb:
             hb.return_value = None
             iKey = '12345678-1234-5678-abcd-12345678abcd'
-            exporter = metrics_exporter.new_metrics_exporter(
+            exporter = new_metrics_exporter(
                 instrumentation_key=iKey, enable_standard_metrics=False)
 
             self.assertEqual(exporter.options.instrumentation_key, iKey)
@@ -239,10 +241,9 @@ class TestAzureMetricsExporter(unittest.TestCase):
                 '.transport.get_exporter_thread')
     def test_new_metrics_exporter_heartbeat(self, exporter_mock):
         with mock.patch('opencensus.ext.azure.metrics_exporter'
-                        '.heartbeat_metrics.enable_heartbeat_metrics') as hb:
+                        '.statsbeat_metrics.collect_statsbeat_metrics') as hb:
             iKey = '12345678-1234-5678-abcd-12345678abcd'
-            exporter = metrics_exporter.new_metrics_exporter(
-                instrumentation_key=iKey)
+            exporter = new_metrics_exporter(instrumentation_key=iKey)
 
             self.assertEqual(exporter.options.instrumentation_key, iKey)
             self.assertEqual(len(hb.call_args_list), 1)